// src/types/fuel/DieselAmmContract.ts
export type IdentityInput = Enum<{ Address: AddressInput, ContractId: ContractIdInput }>;
export type IdentityOutput = Enum<{ Address: AddressOutput, ContractId: ContractIdOutput }>;
// ...
export type StateOutput = Enum<{ Uninitialized: void, Initialized: IdentityOutput, Revoked: void }>;

export type AddressInput = { bits: string };
export type AddressOutput = AddressInput;
export type AssetInput = { id: AssetIdInput, amount: BigNumberish };
export type AssetOutput = { id: AssetIdOutput, amount: BN };
export type AssetIdInput = { bits: string };
export type AssetIdOutput = AssetIdInput;
export type BurnEventInput = { pool_id: [AssetIdInput, AssetIdInput, boolean], recipient: IdentityInput, liquidity: AssetInput, asset_0_out: BigNumberish, asset_1_out: BigNumberish };
export type BurnEventOutput = { pool_id: [AssetIdOutput, AssetIdOutput, boolean], recipient: IdentityOutput, liquidity: AssetOutput, asset_0_out: BN, asset_1_out: BN };
export type ContractIdInput = { bits: string };
export type ContractIdOutput = ContractIdInput;
export type CreatePoolEventInput = { pool_id: [AssetIdInput, AssetIdInput, boolean], decimals_0: BigNumberish, decimals_1: BigNumberish };
export type CreatePoolEventOutput = { pool_id: [AssetIdOutput, AssetIdOutput, boolean], decimals_0: number, decimals_1: number };
export type MintEventInput = { pool_id: [AssetIdInput, AssetIdInput, boolean], recipient: IdentityInput, liquidity: AssetInput, asset_0_in: BigNumberish, asset_1_in: BigNumberish };
export type MintEventOutput = { pool_id: [AssetIdOutput, AssetIdOutput, boolean], recipient: IdentityOutput, liquidity: AssetOutput, asset_0_in: BN, asset_1_in: BN };
// ...
export type PoolMetadataOutput = { reserve_0: BN, reserve_1: BN, liquidity: AssetOutput, decimals_0: number, decimals_1: number };
// ...
export type SwapEventOutput = { pool_id: [AssetIdOutput, AssetIdOutput, boolean], recipient: IdentityOutput, asset_0_in: BN, asset_1_in: BN, asset_0_out: BN, asset_1_out: BN };
// ...
const storageSlots: StorageSlot[] = [];

export class DieselAmmContractInterface extends Interface {
  constructor() {
    super(abi);
  }

  declare functions: {
    owner: FunctionFragment;
    decimals: FunctionFragment;
    name: FunctionFragment;
    symbol: FunctionFragment;
    total_assets: FunctionFragment;
    total_supply: FunctionFragment;
    burn: FunctionFragment;
    create_pool: FunctionFragment;
    fees: FunctionFragment;
    hook: FunctionFragment;
    mint: FunctionFragment;
    pool_metadata: FunctionFragment;
    set_hook: FunctionFragment;
    set_protocol_fees: FunctionFragment;
    swap: FunctionFragment;
    transfer_ownership: FunctionFragment;
  };
}

export class DieselAmmContract extends Contract {
  static readonly abi = abi;
  static readonly storageSlots = storageSlots;

  declare interface: DieselAmmContractInterface;
  declare functions: {
    owner: InvokeFunction<[], StateOutput>;
    decimals: InvokeFunction<[asset: AssetIdInput], Option<number>>;
    name: InvokeFunction<[asset: AssetIdInput], Option<StdString>>;
    symbol: InvokeFunction<[asset: AssetIdInput], Option<StdString>>;
    total_assets: InvokeFunction<[], BN>;
    total_supply: InvokeFunction<[asset: AssetIdInput], Option<BN>>;
    burn: InvokeFunction<[pool_id: [AssetIdInput, AssetIdInput, boolean], to: IdentityInput], [BN, BN]>;
    create_pool: InvokeFunction<[token_0_contract_id: ContractIdInput, token_0_sub_id: string, token_1_contract_id: ContractIdInput, token_1_sub_id: string, is_stable: boolean], [AssetIdOutput, AssetIdOutput, boolean]>;
    fees: InvokeFunction<[], [BN, BN, BN, BN]>;
    hook: InvokeFunction<[], Option<ContractIdOutput>>;
    mint: InvokeFunction<[pool_id: [AssetIdInput, AssetIdInput, boolean], to: IdentityInput], AssetOutput>;
    pool_metadata: InvokeFunction<[pool_id: [AssetIdInput, AssetIdInput, boolean]], Option<PoolMetadataOutput>>;
    set_hook: InvokeFunction<[contract_id?: Option<ContractIdInput>], void>;
    set_protocol_fees: InvokeFunction<[volatile_fee: BigNumberish, stable_fee: BigNumberish], void>;
    swap: InvokeFunction<[pool_id: [AssetIdInput, AssetIdInput, boolean], asset_0_out: BigNumberish, asset_1_out: BigNumberish, to: IdentityInput, data?: Option<Bytes>], void>;
    transfer_ownership: InvokeFunction<[new_owner: IdentityInput], void>;
  };

  constructor(
    id: string | AbstractAddress,
    accountOrProvider: Account | Provider,
  ) {
    super(id, abi, accountOrProvider);
  }
}